requirements.txt
Pillow
langchain-groq
qrcode
pandas
matplotlib
pydantic
langchain
groq
streamlit
rapidfuzz
python-dotenv